requirements.txt
discord.py[voice]
yt-dlp
requests
python-dotenv
PyNaCl